Move ScreenLayout component next to new components
